Way faster and efficient database interactions

+++ Apps/Audiotica.WindowsPhone/PartialView/NowPlayingSheet.xaml.cs
@@ -58,7 +58,7 @@ namespace Audiotica.PartialView
 
             if (song == null) return;
 
-            var currentPlayingId = AppSettingsHelper.Read<long>(PlayerConstants.CurrentTrack);
+            var currentPlayingId = AppSettingsHelper.Read<int>(PlayerConstants.CurrentTrack);
 
             if (e.RemovedItems.Count != 0 && song.Id != currentPlayingId)
             {
